Safari-611.1.10/LayoutTests/accessibility/img-alt-attribute-unassigned-empty-expected.txt





This tests that img elements with an alt attribute and no assigned value, or an empty alt value, are ignored.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS imagesGroup.childrenCount is 2
PASS platformValueForW3CName(imagesGroup.childAtIndex(0)) is "cake0"
PASS platformValueForW3CName(imagesGroup.childAtIndex(1)) is "cake2"
PASS successfullyParsed is true

TEST COMPLETE
